requirements.txt
flask
pillow
supabase
python-dotenv
